pyproject.toml
[project]
name = "streamlit-text-to-image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.70.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "streamlit>=1.44.1",
]
